+++ CHANGES.rst
@@ -1,3 +1,13 @@
+v0.2.0
+======
+- solve may now return non-zero exit (in C++, raises error in Python)
+- check for NaN in solve
+- check for zero diagonal elements in U
+
+v0.1.1
+======
+- LU bindings in Python
+
 v0.1
 ====
 - Initial release with a minimal Python interface